refactor: move hooks module out of settings
This move the hooks class away of settings so it can be used by non-settings
code without import statement looking weird.

It also hides the module as being private.
