+++ ddtrace/settings/__init__.py
@@ -1,7 +1,7 @@
 from .config import Config
 from .exceptions import ConfigException
 from .http import HttpConfig
-from .hooks import Hooks
+from .._hooks import Hooks
 from .integration import IntegrationConfig
 
 # Default global config
